Importing
Adding an already finished project. Was done as part of a school
project with several partners, with the goal being to create a
messaging application that encrypted all messages.

My primary contributions were to the server classes, and the encode and
key classes. I also helped with debugging the other classes.

This is a toy program, do not use it to send sensitive messages as the
encryption is not secure.